Add temporal video workflow and smoke test
