+++ ai-podcast-clipper-backend/temporal/workflows.py
@@ -5,6 +5,7 @@ from temporalio import workflow
 with workflow.unsafe.imports_passed_through():
     from .activities_cpu import ping
     from .activities_gpu import ping_gpu
+    from .activities_video import process_video_activity
 
 @workflow.defn
 class HelloWorkflow:
@@ -27,3 +28,15 @@ class HelloGpuWorkflow:
             task_queue="gpu-tq",
         )
         return f"hello {name}, gpu activity says: {r}"
+
+
+@workflow.defn
+class ProcessVideoWorkflow:
+    @workflow.run
+    async def run(self, s3_key: str) -> dict:
+        result = await workflow.execute_activity(
+            process_video_activity,
+            s3_key,
+            start_to_close_timeout=timedelta(minutes=60),
+        )
+        return result
